requirements.txt
click==7.1.2
unicorn-binance-websocket-api==1.33.0
google-cloud-pubsub==2.8.0